+++ .github/workflows/promptflow-tracing-e2e-test.yml
@@ -28,7 +28,7 @@ jobs:
         name: promptflow-tracing
         path: ${{ env.WORKING_DIRECTORY }}/dist/promptflow_tracing-*.whl
 
-  test:
+  tracing-e2e-test:
     needs: build
     strategy:
       matrix:
@@ -77,7 +77,7 @@ jobs:
             ${{ env.WORKING_DIRECTORY }}/htmlcov/
 
   report:
-    needs: test
+    needs: tracing-e2e-test
     runs-on: ubuntu-latest
     permissions:
       checks: write
