+++ src/promptflow-azure/promptflow/azure/_cli/_connection.py
@@ -7,8 +7,9 @@ from pathlib import Path
 from dotenv import dotenv_values
 
 from promptflow._cli._params import add_param_connection_name, add_param_env, base_params
-from promptflow._cli._utils import _set_workspace_argument_for_subparsers, activate_action, get_client_for_cli
+from promptflow._cli._utils import _set_workspace_argument_for_subparsers, activate_action
 from promptflow._utils.logger_utils import get_cli_sdk_logger
+from promptflow.azure._cli._utils import get_client_for_cli
 from promptflow.connections import CustomConnection
 from promptflow.contracts.types import Secret
