+++ src/promptflow-azure/promptflow/azure/_utils/_tracing.py
@@ -8,11 +8,11 @@ import typing
 from azure.ai.ml import MLClient
 from azure.core.exceptions import ResourceNotFoundError
 
-from promptflow._cli._utils import get_credentials_for_cli
 from promptflow._constants import AzureWorkspaceKind, CosmosDBContainerName
 from promptflow._sdk._utils import extract_workspace_triad_from_trace_provider
 from promptflow._utils.logger_utils import get_cli_sdk_logger
 from promptflow.azure import PFClient
+from promptflow.azure._cli._utils import get_credentials_for_cli
 from promptflow.azure._restclient.flow_service_caller import FlowRequestException
 from promptflow.exceptions import ErrorTarget, UserErrorException
